Consolidate Annotator and core.Annotator
The reasoning behind the original division of these two objects was
flawed. This commit unifies core.Annotator and Annotator into one, which
manages the relationships between all the various components of the
annotator.
